Prepare v0.14.4 release; move internal, top-level functions back to original file (#1389)
